Fix incorrect categories
