+++ package.json
@@ -53,18 +53,18 @@
     "jstat": "^1.9.5"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-env": "^7.18.10",
     "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^22.0.1",
+    "@rollup/plugin-commonjs": "^22.0.2",
     "@rollup/plugin-node-resolve": "^13.3.0",
-    "c8": "^7.11.3",
+    "c8": "^7.12.0",
     "cli-table3": "^0.6.2",
-    "eslint": "^8.19.0",
+    "eslint": "^8.21.0",
     "eslint-config-prettier": "^8.5.0",
     "mocha": "^10.0.0",
     "mochawesome": "^7.1.3",
     "prettier": "^2.7.1",
-    "rollup": "^2.75.7",
+    "rollup": "^2.77.2",
     "rollup-plugin-terser": "^7.0.2",
     "should": "^13.2.3",
     "typescript": "^4.7.4"
