+++ src/database.js
@@ -16,14 +16,7 @@ function compact(array) {
   return result
 }
 
-/**
- * Formula.js only
- *
- * @param {*} database
- * @param {*} title
- * @returns
- */
-export function FINDFIELD(database, title) {
+function findField(database, title) {
   let index = null
 
   utils.arrayEach(database, (value, i) => {
@@ -133,7 +126,7 @@ export function DAVERAGE(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -168,7 +161,7 @@ export function DCOUNT(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -203,7 +196,7 @@ export function DCOUNTA(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -238,7 +231,7 @@ export function DGET(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -278,7 +271,7 @@ export function DMAX(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -315,7 +308,7 @@ export function DMIN(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -352,7 +345,7 @@ export function DPRODUCT(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -394,7 +387,7 @@ export function DSTDEV(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -431,7 +424,7 @@ export function DSTDEVP(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -468,7 +461,7 @@ export function DSUM(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -503,7 +496,7 @@ export function DVAR(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
@@ -539,7 +532,7 @@ export function DVARP(database, field, criteria) {
   let targetFields = []
 
   if (typeof field === 'string') {
-    const index = FINDFIELD(database, field)
+    const index = findField(database, field)
     targetFields = utils.rest(database[index])
   } else {
     targetFields = utils.rest(database[field])
