+++ src/miscellaneous.js
@@ -1,39 +1,6 @@
 import * as error from './utils/error.js'
 import * as utils from './utils/common.js'
 
-/**
- * Returns a list of unique values in a list or range.
- *
- * Category: Lookup and reference
- *
- * @returns
- */
-export function UNIQUE() {
-  const result = []
-
-  for (let i = 0; i < arguments.length; ++i) {
-    let hasElement = false
-    const element = arguments[i]
-
-    // Check if we've already seen this element.
-
-    for (let j = 0; j < result.length; ++j) {
-      hasElement = result[j] === element
-
-      if (hasElement) {
-        break
-      }
-    }
-
-    // If we did not find it, add it to the result.
-    if (!hasElement) {
-      result.push(element)
-    }
-  }
-
-  return result
-}
-
 export const FLATTEN = utils.flatten
 
 export function ARGS2ARRAY() {
