+++ test/logical.js
@@ -20,15 +20,6 @@ describe('Logical', () => {
     logical.AND(0, true).should.equal(false)
   })
 
-  it('CHOOSE', () => {
-    logical.CHOOSE().should.equal(error.na)
-    logical.CHOOSE(1).should.equal(error.na)
-    logical.CHOOSE(1, 'jima').should.equal('jima')
-    logical.CHOOSE(3, 'jima', 'jimb', 'jimc').should.equal('jimc')
-    logical.CHOOSE(2, 'jima').should.equal(error.value)
-    logical.CHOOSE(255, 'jima').should.equal(error.value)
-  })
-
   it('FALSE', () => {
     logical.FALSE().should.equal(false)
   })
