+++ test/miscellaneous.js
@@ -4,13 +4,6 @@ import * as error from '../src/utils/error.js'
 import * as misc from '../src/miscellaneous.js'
 
 describe('Miscellaneous', () => {
-  it('UNIQUE', () => {
-    misc.UNIQUE(1, 2, 3, 4, 5, 6, 6, 3).should.containDeep([1, 2, 3, 4, 5, 6])
-    misc.UNIQUE('jima', 'jimb', 'jima', 'jimc').should.containDeep(['jima', 'jimb', 'jimc'])
-    misc.UNIQUE().should.eql([])
-    misc.UNIQUE([]).should.eql([[]])
-  })
-
   it('ARGS2ARRAY', () => {
     misc.ARGS2ARRAY(1, 2, 3, 4).should.eql([1, 2, 3, 4])
     misc.ARGS2ARRAY('jim', 2, 3.14).should.eql(['jim', 2, 3.14])
